fixing obsoletes versions
